+++ probes/notCreatedInLast90Days/def.yml
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-id: wasCreatedInLast90Days
+id: notCreatedInLast90Days
 short: Checks whether the project was created in the last 90 days.
 motivation: >
   When Scorecard checks the activity of a project in the last 90 days, the project may not have been created before the last 90 days. As such, Scorecard cannot give an accurate score. This probe helps Scorecard assess whether it can give an accurrate score when checking the project activity in the last 90 days.
